+++ src/lib/errors.js
