+++ src/lib/indexing.js
