+++ src/lib/ops.js
